LayoutTests/fast/dom/HTMLTableRowElement/insertCell-skips-non-td-th.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/tabular-data.html#dom-tr-insertcell">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that HTMLTableRowElement.insertCell() skips non &lt;td&gt; / &lt;th&gt; children.");
var tr = document.createElement("tr");
shouldBe("tr.__proto__", "HTMLTableRowElement.prototype");
tr.appendChild(new Text("TEXT"));
tr.appendChild(document.createElement("a"));
tr.insertCell(-1).innerHTML = "1";
// The insertCell() method must create a td element, insert it as a child of the
// tr element, immediately before the indexth td or th element in the cells
// collection, and finally must return the newly created td element.
tr.insertCell(0).innerHTML = "0";
var childNodes = tr.childNodes;
shouldBe("childNodes.length", "4");
shouldBeEqualToString("childNodes[0].nodeValue", "TEXT");
shouldBeEqualToString("childNodes[1].tagName", "A");
shouldBeEqualToString("childNodes[2].innerHTML", "0");
shouldBeEqualToString("childNodes[3].innerHTML", "1");
</script>
</body>
</html>
